Add central exception handling and logging; reorganize PersonServiceImpl class to a different package; update README
