+++ sixth-week-assignment/src/main/java/com/quintrix/java/training/iwterry/springboot/restapi/jparesttemplate/controller/CommentController.java
@@ -17,32 +17,20 @@ public class CommentController {
   private CommentService commentService;
 
   @PostMapping
-  public ResponseEntity<?> createComment(@RequestBody CommentDto commentDto) {
-    try {
-      CommentDto createdCommentDto = commentService.addComment(commentDto);
-      return new ResponseEntity<>(createdCommentDto, HttpStatus.CREATED);
-    } catch (HttpStatusCodeException e) {
-      return new ResponseEntity<>(e.getMessage(), e.getStatusCode());
-    }
+  public ResponseEntity<CommentDto> createComment(@RequestBody CommentDto commentDto) {
+    CommentDto createdCommentDto = commentService.addComment(commentDto);
+    return new ResponseEntity<>(createdCommentDto, HttpStatus.CREATED);
   }
 
   @GetMapping
-  public ResponseEntity<?> getComments(@RequestParam(name = "name", defaultValue = "") String name) {
-    try {
-      List<CommentDto> commentDtos = commentService.getCommentsContainingName(name);
-      return new ResponseEntity<>(commentDtos, HttpStatus.OK);
-    } catch (HttpStatusCodeException e) {
-      return new ResponseEntity<>(e.getMessage(), e.getStatusCode());
-    }
+  public ResponseEntity<List<CommentDto>> getComments(@RequestParam(name = "name", defaultValue = "") String name) {
+    List<CommentDto> commentDtos = commentService.getCommentsContainingName(name);
+    return new ResponseEntity<>(commentDtos, HttpStatus.OK);
   }
 
   @GetMapping("/{id}")
-  public ResponseEntity<?> getCommentById(@PathVariable(name = "id") long id) {
-    try {
-      CommentDto commentDto = commentService.getCommentById(id);
-      return new ResponseEntity<>(commentDto, HttpStatus.OK);
-    } catch (HttpStatusCodeException e) {
-      return new ResponseEntity<>(e.getMessage(), e.getStatusCode());
-    }
+  public ResponseEntity<CommentDto> getCommentById(@PathVariable(name = "id") long id) {
+    CommentDto commentDto = commentService.getCommentById(id);
+    return new ResponseEntity<>(commentDto, HttpStatus.OK);
   }
 }
