+++ sixth-week-assignment/src/main/resources/logback.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<configuration debug="true">
+    <appender name="stdout" class="ch.qos.logback.core.ConsoleAppender">
+        <encoder>
+            <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
+        </encoder>
+    </appender>
+
+    <appender name="file" class="ch.qos.logback.core.FileAppender">
+        <file>logs/error.log</file>
+        <append>true</append>
+        <encoder>
+            <pattern>%d{yyyy-MM-dd} at %d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
+        </encoder>
+    </appender>
+
+    <!-- As for practice, I am using different logging levels for different packages. -->
+    <logger name="com.quintrix.java.training.iwterry.springboot.restapi.jparesttemplate.exception" level="error">
+        <appender-ref ref="file" />
+    </logger>
+    <logger name="com.quintrix.java.training.iwterry.springboot.restapi.jparesttemplate.httprequestservice" level="debug" />
+
+    <root level="info">
+        <appender-ref ref="stdout" />
+    </root>
+</configuration>
